Sync between working branches
Just syncing between two other working branches not to loose code.
Removed initial css syntax and app should be in working condition atm.
