+++ src/app/Components/Header/Header.js
@@ -1,0 +1,35 @@
+import * as React from "react";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import Link from "@mui/material/Link";
+import styled from "styled-components";
+
+function Header(props) {
+  const { sections, title } = props;
+
+  return (
+    <React.Fragment>
+      <Toolbar sx={{ borderBottom: 1, borderColor: "divider" }}></Toolbar>
+      <Toolbar
+        component="nav"
+        variant="dense"
+        sx={{ justifyContent: "space-between", overflowX: "auto" }}
+      >
+        {sections.map((section) => (
+          <Link
+            color="inherit"
+            noWrap
+            key={section.title}
+            variant="body2"
+            href={section.url}
+            sx={{ p: 1, flexShrink: 0 }}
+          >
+            {section.title}
+          </Link>
+        ))}
+      </Toolbar>
+    </React.Fragment>
+  );
+}
+
+export default Header;
